refactor(rippling): extract pure JS helpers to composables/rippling/
First step of breaking RipplingExplorer.vue into subcomponents.  Pulls
the eight pure-function helpers (chaikinSmooth, geoToLeaflet,
crowFliesKm, pointInRing, segmentsIntersect, classifyPost, thumbUrlFor,
formatTimeAgo, escapeHTML, renderPie) out of the 2900-line monolith
into three small modules:

- composables/rippling/geometry.js — Chaikin smoothing, haversine, ring
  containment, segment intersection.
- composables/rippling/scoring.js  — post classification, thumbnail URL
  resolution, relative-time formatting, HTML-escape.
- composables/rippling/pie.js      — SVG pie-chart renderer.

Zero behaviour change — verified via HMR (no parse errors) and a
visual smoke test of the rippling page.
